+++ src/hash.rs
@@ -1,5 +1,4 @@
 use crate::structs::{HashParts, StringUtils, AlgoType};
-use crate::base64;
 
 /// 22 character salt + 31 character hash
 fn valid_salt_and_hash(salt_and_hash: &str) -> bool {
@@ -72,14 +71,6 @@ pub fn get_salt(hash: &str) -> String {
 /// Get the hashed password in the given hash
 pub fn get_hash(hash: &str) -> String {
     split_hash_into_parts(hash).unwrap().hash()
-}
-
-pub fn get_salt_bytes(hash: &str) -> Vec<u8> {
-    base64::decode_bcrypt(split_hash_into_parts(hash).unwrap().salt())
-}
-
-pub fn get_hash_bytes(hash: &str) -> Vec<u8> {
-    base64::decode_bcrypt(split_hash_into_parts(hash).unwrap().hash())
 }
 
 #[cfg(test)]