+++ src/lib.rs
@@ -1,6 +1,5 @@
 mod hash;
 mod structs;
-mod base64;
 
 use wasm_bindgen::prelude::*;
 
@@ -22,14 +21,4 @@ pub fn get_salt(hash: &str) -> String {
 #[wasm_bindgen]
 pub fn get_hash(hash: &str) -> String {
     hash::get_hash(hash)
-}
-
-#[wasm_bindgen]
-pub fn get_salt_bytes(hash: &str) -> Vec<u8> {
-    hash::get_salt_bytes(hash)
-}
-
-#[wasm_bindgen]
-pub fn get_hash_bytes(hash: &str) -> Vec<u8> {
-    hash::get_hash_bytes(hash)
 }